+++ src/storage/mod.rs
@@ -10,10 +10,16 @@
 //! - Async trait methods handle I/O operations with the database
 //! - All storage operations use existing domain types for type safety
 //!
+//! # Submodules
+//!
+//! - [`agent_storage`](agent_storage) - SQLite implementation for agent registry persistence
+//! - [`message_storage`](message_storage) - SQLite implementation for FIPA message persistence
+//! - [`conversation_storage`](conversation_storage) - SQLite implementation for conversation state persistence
+//!
 //! # Usage
 //!
 //! ```rust,ignore
-//! use crate::storage::AgentStorage;
+//! use crate::storage::{AgentStorage, SqliteAgentStorage};
 //! use crate::domain_types::{AgentId, AgentName};
 //!
 //! async fn persist_agent(storage: &dyn AgentStorage, id: AgentId, name: AgentName) {
@@ -22,26 +28,10 @@
 //! ```
 
 use async_trait::async_trait;
-use sqlx::Row;
-use std::{
-    borrow::Cow,
-    collections::HashSet,
-    fmt::Write,
-    io::{Error as IoError, ErrorKind},
-    time::{Duration, SystemTime, UNIX_EPOCH},
-};
-use tracing::{info, instrument, warn};
 
 use crate::{
-    database::{DatabaseConnection, DatabaseResult},
+    database::DatabaseResult,
     domain_types::{AgentId, AgentName},
-    message_router::{
-        domain_types::{
-            Conversation, ConversationCreatedAt, ConversationId, FipaMessage, MessageContent,
-            MessageCount, MessageId, MessageTimestamp, Performative, ProtocolName,
-        },
-        traits::{ConversationError, RouterError},
-    },
 };
 
 /// Persistent storage interface for agent registry operations.
@@ -128,858 +118,15 @@ pub trait AgentStorage: Send + Sync {
     async fn remove_agent(&self, agent_id: AgentId) -> DatabaseResult<()>;
 }
 
-/// `SQLite` implementation of message storage for FIPA message persistence.
-///
-/// This implementation stores FIPA messages in a dedicated `message_storage` table
-/// created via the migration system (`003_create_message_storage.sql`). The table
-/// schema includes optimized indexes for common query patterns including sender/receiver
-/// lookups and conversation-based filtering.
-///
-/// # Migration System Integration
-///
-/// Table creation is handled by the migration system during `DatabaseConnection::initialize()`.
-/// This ensures consistent schema versioning and eliminates the need for CREATE TABLE
-/// IF NOT EXISTS patterns. The migration includes performance indexes for:
-///
-/// - `sender_id` lookups with temporal ordering
-/// - `receiver_id` lookups with temporal ordering
-/// - `conversation_id` filtering with temporal ordering
-/// - `created_at` temporal queries
-///
-/// # Performance Characteristics
-///
-/// - Target: < 1ms average query time for individual message operations
-/// - Optimized indexes support efficient message routing and conversation retrieval
-/// - UUID format validation via CHECK constraints ensures data integrity
-/// - Schema designed for high-throughput message processing
-///
-/// # Implementation Status
-///
-/// Currently provides placeholder methods returning `unimplemented!()` errors.
-/// Full implementation with proper domain type integration, error handling,
-/// and observability will be completed in future development cycles.
-pub struct SqliteMessageStorage {
-    _connection: DatabaseConnection,
-}
-
-impl SqliteMessageStorage {
-    /// Creates a new `SqliteMessageStorage` instance.
-    ///
-    /// The provided database connection should already be initialized with the migration
-    /// system, ensuring the `message_storage` table and indexes are available.
-    ///
-    /// # Arguments
-    ///
-    /// * `connection` - Database connection with migration-created schema
-    ///
-    /// # Returns
-    ///
-    /// A new message storage instance ready for FIPA message persistence operations.
-    pub fn new(connection: DatabaseConnection) -> Self {
-        Self {
-            _connection: connection,
-        }
-    }
-
-    /// Parses a database row into a `FipaMessage`.
-    ///
-    /// This helper method reconstructs `FipaMessage` instances from `SQLite` row data,
-    /// eliminating duplication between `get_message` and `list_agent_messages` operations.
-    fn parse_message_from_row(row: &sqlx::sqlite::SqliteRow) -> Result<FipaMessage, RouterError> {
-        let message_id_str: String = row.get("message_id");
-        let (message_id, sender, receiver) = Self::parse_message_identifiers(row, &message_id_str)?;
-        let conversation_id = Self::parse_conversation_id(row)?;
-        let message_content = Self::parse_message_content_from_row(row, &message_id_str)?;
-        let performative = Self::parse_performative_from_row(row)?;
-        let created_at = Self::parse_timestamp_from_row(row, &message_id_str)?;
-
-        Ok(FipaMessage {
-            performative,
-            sender,
-            receiver,
-            content: message_content,
-            language: None, // Not stored in current schema
-            ontology: None, // Not stored in current schema
-            protocol: None, // Not stored in current schema
-            conversation_id,
-            reply_with: None,  // Not stored in current schema
-            in_reply_to: None, // Not stored in current schema
-            message_id,
-            created_at,
-            trace_context: None, // Not stored in current schema
-            delivery_options: crate::message_router::domain_types::DeliveryOptions::default(),
-        })
-    }
-
-    /// Parses message identifiers (`message_id`, `sender_id`, `receiver_id`) from database row.
-    fn parse_message_identifiers(
-        row: &sqlx::sqlite::SqliteRow,
-        message_id_str: &str,
-    ) -> Result<(MessageId, AgentId, AgentId), RouterError> {
-        let sender_str: String = row.get("sender_id");
-        let receiver_str: String = row.get("receiver_id");
-
-        let parsed_message_id =
-            uuid::Uuid::parse_str(message_id_str).map_err(|e| RouterError::StorageError {
-                source: Box::new(IoError::new(
-                    ErrorKind::InvalidData,
-                    format!("Invalid message ID UUID '{message_id_str}': {e}"),
-                )),
-            })?;
-
-        let parsed_sender =
-            uuid::Uuid::parse_str(&sender_str).map_err(|e| RouterError::StorageError {
-                source: Box::new(IoError::new(
-                    ErrorKind::InvalidData,
-                    format!("Invalid sender UUID '{sender_str}': {e}"),
-                )),
-            })?;
-
-        let parsed_receiver =
-            uuid::Uuid::parse_str(&receiver_str).map_err(|e| RouterError::StorageError {
-                source: Box::new(IoError::new(
-                    ErrorKind::InvalidData,
-                    format!("Invalid receiver UUID '{receiver_str}': {e}"),
-                )),
-            })?;
-
-        Ok((
-            MessageId::new(parsed_message_id),
-            AgentId::new(parsed_sender),
-            AgentId::new(parsed_receiver),
-        ))
-    }
-
-    /// Parses optional conversation ID from database row.
-    fn parse_conversation_id(
-        row: &sqlx::sqlite::SqliteRow,
-    ) -> Result<Option<ConversationId>, RouterError> {
-        let conversation_id_str: Option<String> = row.get("conversation_id");
-        if let Some(conv_str) = conversation_id_str {
-            let parsed_uuid =
-                uuid::Uuid::parse_str(&conv_str).map_err(|e| RouterError::StorageError {
-                    source: Box::new(IoError::new(
-                        ErrorKind::InvalidData,
-                        format!("Invalid conversation UUID '{conv_str}': {e}"),
-                    )),
-                })?;
-            Ok(Some(ConversationId::new(parsed_uuid)))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Parses message content from database row.
-    fn parse_message_content_from_row(
-        row: &sqlx::sqlite::SqliteRow,
-        message_id_str: &str,
-    ) -> Result<MessageContent, RouterError> {
-        let content_string: String = row.get("message_content");
-        let content_bytes = Self::parse_message_content(&content_string, message_id_str)?;
-        MessageContent::try_new(content_bytes.into_owned()).map_err(|e| RouterError::StorageError {
-            source: Box::new(IoError::new(
-                ErrorKind::InvalidData,
-                format!("Invalid message content for message '{message_id_str}': {e}"),
-            )),
-        })
-    }
-
-    /// Parses performative from database row.
-    fn parse_performative_from_row(
-        row: &sqlx::sqlite::SqliteRow,
-    ) -> Result<Performative, RouterError> {
-        let performative_str: String = row.get("performative");
-        Self::parse_performative(&performative_str)
-    }
-
-    /// Parses timestamp from database row.
-    fn parse_timestamp_from_row(
-        row: &sqlx::sqlite::SqliteRow,
-        message_id_str: &str,
-    ) -> Result<MessageTimestamp, RouterError> {
-        let created_at_secs: i64 = row.get("created_at");
-        Self::parse_timestamp(created_at_secs, message_id_str)
-    }
-
-    /// Parses length-prefixed message content in format: `{length}::{content}`
-    fn parse_message_content<'a>(
-        content_string: &'a str,
-        message_id: &str,
-    ) -> Result<Cow<'a, [u8]>, RouterError> {
-        let parts: Vec<&str> = content_string.splitn(2, "::").collect();
-        if parts.len() != 2 {
-            return Err(RouterError::StorageError {
-                source: Box::new(IoError::new(
-                    ErrorKind::InvalidData,
-                    format!(
-                        "Invalid length-prefixed content format for message '{message_id}': expected 'length::content', got '{content_string}'"
-                    ),
-                )),
-            });
-        }
-        Ok(Cow::Borrowed(parts[1].as_bytes()))
-    }
-
-    /// Parses performative string into enum variant.
-    fn parse_performative(performative_str: &str) -> Result<Performative, RouterError> {
-        match performative_str {
-            "Inform" => Ok(Performative::Inform),
-            "Request" => Ok(Performative::Request),
-            "Agree" => Ok(Performative::Agree),
-            "Refuse" => Ok(Performative::Refuse),
-            "NotUnderstood" => Ok(Performative::NotUnderstood),
-            "Failure" => Ok(Performative::Failure),
-            "QueryIf" => Ok(Performative::QueryIf),
-            "QueryRef" => Ok(Performative::QueryRef),
-            "Propose" => Ok(Performative::Propose),
-            "AcceptProposal" => Ok(Performative::AcceptProposal),
-            "RejectProposal" => Ok(Performative::RejectProposal),
-            "Heartbeat" => Ok(Performative::Heartbeat),
-            "Capability" => Ok(Performative::Capability),
-            unknown => Err(RouterError::StorageError {
-                source: Box::new(IoError::new(
-                    ErrorKind::InvalidData,
-                    format!(
-                        "Unknown performative '{unknown}'. Valid values: Inform, Request, Agree, Refuse, NotUnderstood, Failure, QueryIf, QueryRef, Propose, AcceptProposal, RejectProposal, Heartbeat, Capability"
-                    ),
-                )),
-            }),
-        }
-    }
-
-    /// Serializes message content to length-prefixed format.
-    fn serialize_message_content(content_bytes: &[u8]) -> String {
-        let content_len = content_bytes.len();
-        let mut length_digits = 1;
-        let mut temp = content_len;
-        while temp >= 10 {
-            length_digits += 1;
-            temp /= 10;
-        }
-        let estimated_capacity = length_digits + 2 + content_len;
-
-        let mut result = String::with_capacity(estimated_capacity);
-        write!(&mut result, "{content_len}::").expect("Writing length should never fail");
-
-        if let Ok(valid_utf8) = std::str::from_utf8(content_bytes) {
-            result.push_str(valid_utf8);
-        } else {
-            let content_string = String::from_utf8_lossy(content_bytes);
-            result.push_str(&content_string);
-        }
-
-        result
-    }
-
-    /// Converts performative to string representation.
-    fn performative_to_str(performative: Performative) -> &'static str {
-        match performative {
-            Performative::Request => "Request",
-            Performative::Inform => "Inform",
-            Performative::Agree => "Agree",
-            Performative::Refuse => "Refuse",
-            Performative::NotUnderstood => "NotUnderstood",
-            Performative::Failure => "Failure",
-            Performative::QueryIf => "QueryIf",
-            Performative::QueryRef => "QueryRef",
-            Performative::Propose => "Propose",
-            Performative::AcceptProposal => "AcceptProposal",
-            Performative::RejectProposal => "RejectProposal",
-            Performative::Heartbeat => "Heartbeat",
-            Performative::Capability => "Capability",
-        }
-    }
-
-    /// Converts `MessageTimestamp` to Unix seconds using standard library.
-    fn timestamp_to_unix_secs(timestamp: MessageTimestamp) -> Result<i64, RouterError> {
-        let duration = timestamp
-            .into_inner()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default();
-        i64::try_from(duration.as_secs()).map_err(|e| RouterError::StorageError {
-            source: Box::new(e),
-        })
-    }
-
-    /// Parses Unix timestamp using standard library duration arithmetic.
-    fn parse_timestamp(
-        created_at_secs: i64,
-        message_id: &str,
-    ) -> Result<MessageTimestamp, RouterError> {
-        let timestamp_u64 =
-            u64::try_from(created_at_secs).map_err(|e| RouterError::StorageError {
-                source: Box::new(IoError::new(
-                    ErrorKind::InvalidData,
-                    format!("Invalid timestamp {created_at_secs} for message '{message_id}': {e}"),
-                )),
-            })?;
-
-        Ok(MessageTimestamp::new(
-            UNIX_EPOCH + Duration::from_secs(timestamp_u64),
-        ))
-    }
-}
-
-#[async_trait]
-impl crate::message_router::traits::MessageStorage for SqliteMessageStorage {
-    #[instrument(skip(self, message), fields(message_id = %message.message_id))]
-    async fn store_message(&self, message: &FipaMessage) -> Result<(), RouterError> {
-        let message_content_string = Self::serialize_message_content(message.content.as_bytes());
-        let performative_str = Self::performative_to_str(message.performative);
-        let created_at = Self::timestamp_to_unix_secs(message.created_at)?;
-
-        sqlx::query(INSERT_MESSAGE)
-            .bind(message.message_id.to_string())
-            .bind(message.sender.to_string())
-            .bind(message.receiver.to_string())
-            .bind(
-                message
-                    .conversation_id
-                    .as_ref()
-                    .map(std::string::ToString::to_string),
-            )
-            .bind(message_content_string)
-            .bind(performative_str)
-            .bind(created_at)
-            .bind(Some(created_at + 2)) // 2 second TTL
-            .execute(self._connection.pool())
-            .await
-            .map_err(|e| {
-                warn!(
-                    "Failed to store message {} from {} to {}: {}",
-                    message.message_id, message.sender, message.receiver, e
-                );
-                RouterError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-
-        info!("Stored message {}", message.message_id);
-        Ok(())
-    }
-
-    #[instrument(skip(self), fields(message_id = %message_id))]
-    async fn get_message(&self, message_id: MessageId) -> Result<Option<FipaMessage>, RouterError> {
-        let row = sqlx::query(SELECT_MESSAGE_BY_ID)
-            .bind(message_id.to_string())
-            .fetch_optional(self._connection.pool())
-            .await
-            .map_err(|e| {
-                warn!("Failed to retrieve message {}: {}", message_id, e);
-                RouterError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-
-        let Some(row) = row else {
-            info!("Message {} not found in storage", message_id);
-            return Ok(None);
-        };
-
-        // Use centralized message parsing logic
-        let fipa_message = Self::parse_message_from_row(&row)?;
-
-        info!(
-            "Retrieved message {} from sender {} to receiver {}",
-            message_id, fipa_message.sender, fipa_message.receiver
-        );
-
-        Ok(Some(fipa_message))
-    }
-
-    #[instrument(skip(self), fields(message_id = %message_id))]
-    async fn remove_message(&self, message_id: MessageId) -> Result<(), RouterError> {
-        let result = sqlx::query(DELETE_MESSAGE_BY_ID)
-            .bind(message_id.to_string())
-            .execute(self._connection.pool())
-            .await
-            .map_err(|e| {
-                warn!("Failed to remove message {}: {}", message_id, e);
-                RouterError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-
-        let rows_affected = result.rows_affected();
-        if rows_affected > 0 {
-            info!(
-                "Removed message {} ({} rows affected)",
-                message_id, rows_affected
-            );
-        } else {
-            info!(
-                "Message {} not found for removal (idempotent operation)",
-                message_id
-            );
-        }
-
-        Ok(())
-    }
-
-    #[instrument(skip(self), fields(agent_id = %agent_id, limit = ?limit))]
-    async fn list_agent_messages(
-        &self,
-        agent_id: AgentId,
-        limit: Option<usize>,
-    ) -> Result<Vec<FipaMessage>, RouterError> {
-        // Choose appropriate query based on limit parameter to avoid dynamic SQL construction
-        let rows = if let Some(limit_value) = limit {
-            // Use safe conversion with reasonable maximum limit to prevent memory issues
-            let limit_i64 = i64::try_from(limit_value).unwrap_or(1000);
-
-            sqlx::query(SELECT_MESSAGES_FOR_AGENT_LIMITED)
-                .bind(agent_id.to_string())
-                .bind(limit_i64)
-                .fetch_all(self._connection.pool())
-                .await
-                .map_err(|e| {
-                    warn!(
-                        "Failed to list messages for agent {} (limited to {}): {}",
-                        agent_id, limit_value, e
-                    );
-                    crate::message_router::traits::RouterError::StorageError {
-                        source: Box::new(e),
-                    }
-                })?
-        } else {
-            sqlx::query(SELECT_MESSAGES_FOR_AGENT)
-                .bind(agent_id.to_string())
-                .fetch_all(self._connection.pool())
-                .await
-                .map_err(|e| {
-                    warn!("Failed to list all messages for agent {}: {}", agent_id, e);
-                    crate::message_router::traits::RouterError::StorageError {
-                        source: Box::new(e),
-                    }
-                })?
-        };
-
-        // Use centralized message parsing for consistency and maintainability
-        let mut messages = Vec::with_capacity(rows.len());
-        for row in &rows {
-            let fipa_message = Self::parse_message_from_row(row)?;
-            messages.push(fipa_message);
-        }
-
-        info!(
-            "Listed {} messages for agent {} {}",
-            messages.len(),
-            agent_id,
-            if let Some(limit_value) = limit {
-                format!("(limited to {limit_value})")
-            } else {
-                "(no limit)".to_string()
-            }
-        );
-
-        Ok(messages)
-    }
-}
-
-// SQL Constants for ConversationStorage
-// Table creation handled by migration system - see migrations/004_create_conversations.sql and migrations/005_create_conversation_participants.sql
-
-const INSERT_CONVERSATION: &str = r"
-INSERT OR REPLACE INTO conversations (
-    conversation_id, protocol_name, created_at, last_activity, message_count, is_archived
-) VALUES (?, ?, ?, ?, ?, ?);
-";
-
-const INSERT_PARTICIPANT: &str = r"
-INSERT INTO conversation_participants (conversation_id, participant_id)
-VALUES (?, ?);
-";
-
-const DELETE_PARTICIPANTS: &str = r"
-DELETE FROM conversation_participants WHERE conversation_id = ?;
-";
-
-const SELECT_CONVERSATION: &str = r"
-SELECT conversation_id, protocol_name, created_at, last_activity, message_count, is_archived
-FROM conversations WHERE conversation_id = ?;
-";
-
-const SELECT_PARTICIPANTS: &str = r"
-SELECT participant_id FROM conversation_participants WHERE conversation_id = ?;
-";
-
-const UPDATE_CONVERSATION_ARCHIVE_STATUS: &str = r"
-UPDATE conversations SET is_archived = TRUE WHERE conversation_id = ?;
-";
-
-const SELECT_AGENT_CONVERSATIONS: &str = r"
-SELECT DISTINCT c.conversation_id
-FROM conversations c
-JOIN conversation_participants p ON c.conversation_id = p.conversation_id
-WHERE p.participant_id = ? AND c.is_archived = FALSE
-ORDER BY c.last_activity DESC;
-";
-
-const INSERT_MESSAGE: &str = r"
-INSERT OR REPLACE INTO message_storage (
-    message_id, sender_id, receiver_id, conversation_id, message_content, performative, created_at, expires_at
-) VALUES (?, ?, ?, ?, ?, ?, ?, ?);
-";
-
-const SELECT_MESSAGE_BY_ID: &str = r"
-SELECT message_id, sender_id, receiver_id, conversation_id, message_content, performative, created_at, expires_at
-FROM message_storage
-WHERE message_id = ? AND (expires_at IS NULL OR expires_at > strftime('%s', 'now'));
-";
-
-const DELETE_MESSAGE_BY_ID: &str = r"
-DELETE FROM message_storage WHERE message_id = ?;
-";
-
-const SELECT_MESSAGES_FOR_AGENT: &str = r"
-SELECT message_id, sender_id, receiver_id, conversation_id, message_content, performative, created_at, expires_at
-FROM message_storage
-WHERE sender_id = ?1 OR receiver_id = ?1
-ORDER BY created_at DESC;
-";
-
-const SELECT_MESSAGES_FOR_AGENT_LIMITED: &str = r"
-SELECT message_id, sender_id, receiver_id, conversation_id, message_content, performative, created_at, expires_at
-FROM message_storage
-WHERE sender_id = ?1 OR receiver_id = ?1
-ORDER BY created_at DESC
-LIMIT ?2;
-";
-
-/// `SQLite` implementation of conversation storage for FIPA conversation persistence.
-///
-/// This implementation stores conversation state and participants in dedicated
-/// `conversations` and `conversation_participants` tables created via the migration
-/// system (`004_create_conversations.sql`, `005_create_conversation_participants.sql`).
-/// The tables include optimized indexes for common query patterns.
-///
-/// # Migration System Integration
-///
-/// Table creation is handled by the migration system during `DatabaseConnection::initialize()`.
-/// This ensures consistent schema versioning and eliminates the need for CREATE TABLE
-/// IF NOT EXISTS patterns. The migration includes performance indexes for:
-///
-/// - conversation archival status filtering
-/// - temporal ordering for last activity queries
-/// - protocol-based conversation filtering
-/// - efficient participant lookups
-///
-/// # Performance Characteristics
-///
-/// - Target: < 1ms average query time for individual conversation operations
-/// - Optimized indexes support efficient conversation and participant queries
-/// - Foreign key constraints ensure referential integrity
-/// - Schema designed for multi-participant conversation scenarios
-pub struct SqliteConversationStorage {
-    connection: DatabaseConnection,
-}
-
-impl SqliteConversationStorage {
-    /// Creates a new `SqliteConversationStorage` instance.
-    ///
-    /// The provided database connection should already be initialized with the migration
-    /// system, ensuring the `conversations` and `conversation_participants` tables and indexes
-    /// are available.
-    ///
-    /// # Arguments
-    ///
-    /// * `connection` - Database connection with migration-created schema
-    ///
-    /// # Returns
-    ///
-    /// A new conversation storage instance ready for FIPA conversation persistence operations.
-    pub fn new(connection: DatabaseConnection) -> Self {
-        Self { connection }
-    }
-
-    /// Converts `SystemTime` to Unix seconds using standard library.
-    fn system_time_to_unix_secs(time: SystemTime) -> Result<i64, ConversationError> {
-        let duration = time.duration_since(UNIX_EPOCH).unwrap_or_default();
-        i64::try_from(duration.as_secs()).map_err(|e| ConversationError::StorageError {
-            source: Box::new(e),
-        })
-    }
-}
-
-#[async_trait]
-impl crate::message_router::traits::ConversationStorage for SqliteConversationStorage {
-    #[instrument(skip(self, conversation))]
-    async fn save_conversation(
-        &self,
-        conversation: &Conversation,
-    ) -> Result<(), ConversationError> {
-        // Schema is already initialized via migration system during DatabaseConnection::initialize()
-        let pool = self.connection.pool();
-
-        // Clear existing participants
-        sqlx::query(DELETE_PARTICIPANTS)
-            .bind(conversation.id.to_string())
-            .execute(pool)
-            .await
-            .map_err(|e| {
-                warn!(
-                    "Failed to delete participants for conversation {}: {}",
-                    conversation.id, e
-                );
-                ConversationError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-
-        // Insert conversation record
-        let protocol_name = conversation
-            .protocol
-            .as_ref()
-            .map(std::string::ToString::to_string);
-        let created_at = Self::system_time_to_unix_secs(conversation.created_at.into_inner())
-            .map_err(|e| {
-                warn!(
-                    "Failed to convert created_at timestamp for conversation {}: {}",
-                    conversation.id, e
-                );
-                e
-            })?;
-        let last_activity = Self::system_time_to_unix_secs(conversation.last_activity.into_inner())
-            .map_err(|e| {
-                warn!(
-                    "Failed to convert last_activity timestamp for conversation {}: {}",
-                    conversation.id, e
-                );
-                e
-            })?;
-        let message_count =
-            u32::try_from(conversation.message_count.into_inner()).map_err(|e| {
-                warn!(
-                    "Failed to convert message_count for conversation {}: {}",
-                    conversation.id, e
-                );
-                ConversationError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-
-        sqlx::query(INSERT_CONVERSATION)
-            .bind(conversation.id.to_string())
-            .bind(protocol_name)
-            .bind(created_at)
-            .bind(last_activity)
-            .bind(message_count)
-            .bind(false) // is_archived
-            .execute(pool)
-            .await
-            .map_err(|e| {
-                warn!("Failed to insert conversation {}: {}", conversation.id, e);
-                ConversationError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-
-        // Insert participants
-        for participant in &conversation.participants {
-            sqlx::query(INSERT_PARTICIPANT)
-                .bind(conversation.id.to_string())
-                .bind(participant.to_string())
-                .execute(pool)
-                .await
-                .map_err(|e| {
-                    warn!(
-                        "Failed to insert participant {} for conversation {}: {}",
-                        participant, conversation.id, e
-                    );
-                    ConversationError::StorageError {
-                        source: Box::new(e),
-                    }
-                })?;
-        }
-
-        info!(
-            "Saved conversation {} with {} participants",
-            conversation.id,
-            conversation.participants.len()
-        );
-        Ok(())
-    }
-
-    #[instrument(skip(self))]
-    async fn load_conversation(
-        &self,
-        conversation_id: ConversationId,
-    ) -> Result<Option<Conversation>, ConversationError> {
-        let pool = self.connection.pool();
-
-        // Load conversation record
-        let conversation_row = sqlx::query(SELECT_CONVERSATION)
-            .bind(conversation_id.to_string())
-            .fetch_optional(pool)
-            .await
-            .map_err(|e| {
-                warn!("Failed to load conversation {}: {}", conversation_id, e);
-                ConversationError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-
-        let Some(conversation_row) = conversation_row else {
-            return Ok(None);
-        };
-
-        // Load participants
-        let participant_rows = sqlx::query(SELECT_PARTICIPANTS)
-            .bind(conversation_id.to_string())
-            .fetch_all(pool)
-            .await
-            .map_err(|e| {
-                warn!(
-                    "Failed to load participants for conversation {}: {}",
-                    conversation_id, e
-                );
-                ConversationError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-
-        // Parse participants
-        let mut participants = HashSet::new();
-        for row in participant_rows {
-            let participant_str: String = row.get("participant_id");
-            let participant_uuid = participant_str.parse::<uuid::Uuid>().map_err(|e| {
-                warn!(
-                    "Failed to parse participant UUID {}: {}",
-                    participant_str, e
-                );
-                ConversationError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-            participants.insert(AgentId::new(participant_uuid));
-        }
-
-        // Parse conversation fields
-        let protocol_name_str: Option<String> = conversation_row.get("protocol_name");
-        let protocol = if let Some(name) = protocol_name_str {
-            Some(
-                ProtocolName::try_new(name).map_err(|e| ConversationError::StorageError {
-                    source: Box::new(e),
-                })?,
-            )
-        } else {
-            None
-        };
-
-        let created_at_secs: i64 = conversation_row.get("created_at");
-        let created_at = ConversationCreatedAt::new(
-            UNIX_EPOCH
-                + Duration::from_secs(u64::try_from(created_at_secs).map_err(|e| {
-                    warn!(
-                        "Failed to convert created_at timestamp for conversation {}: {}",
-                        conversation_id, e
-                    );
-                    ConversationError::StorageError {
-                        source: Box::new(e),
-                    }
-                })?),
-        );
-
-        let last_activity_secs: i64 = conversation_row.get("last_activity");
-        let last_activity = MessageTimestamp::new(
-            UNIX_EPOCH
-                + Duration::from_secs(u64::try_from(last_activity_secs).map_err(|e| {
-                    warn!(
-                        "Failed to convert last_activity timestamp for conversation {}: {}",
-                        conversation_id, e
-                    );
-                    ConversationError::StorageError {
-                        source: Box::new(e),
-                    }
-                })?),
-        );
-
-        let message_count_val: u32 = conversation_row.get("message_count");
-        let message_count = MessageCount::new(message_count_val as usize);
-
-        let conversation = Conversation {
-            id: conversation_id,
-            participants,
-            protocol,
-            created_at,
-            last_activity,
-            message_count,
-        };
-
-        info!(
-            "Loaded conversation {} with {} participants",
-            conversation_id,
-            conversation.participants.len()
-        );
-        Ok(Some(conversation))
-    }
-
-    #[instrument(skip(self, conversation))]
-    async fn archive_conversation(
-        &self,
-        conversation: &Conversation,
-    ) -> Result<(), ConversationError> {
-        let pool = self.connection.pool();
-
-        sqlx::query(UPDATE_CONVERSATION_ARCHIVE_STATUS)
-            .bind(conversation.id.to_string())
-            .execute(pool)
-            .await
-            .map_err(|e| {
-                warn!("Failed to archive conversation {}: {}", conversation.id, e);
-                ConversationError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-
-        info!("Archived conversation {}", conversation.id);
-        Ok(())
-    }
-
-    #[instrument(skip(self))]
-    async fn list_agent_conversations(
-        &self,
-        agent_id: AgentId,
-    ) -> Result<Vec<ConversationId>, ConversationError> {
-        let pool = self.connection.pool();
-
-        let rows = sqlx::query(SELECT_AGENT_CONVERSATIONS)
-            .bind(agent_id.to_string())
-            .fetch_all(pool)
-            .await
-            .map_err(|e| {
-                warn!("Failed to list conversations for agent {}: {}", agent_id, e);
-                ConversationError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-
-        let mut conversation_ids = Vec::new();
-        for row in rows {
-            let conversation_id_str: String = row.get("conversation_id");
-            let conversation_uuid = conversation_id_str.parse::<uuid::Uuid>().map_err(|e| {
-                warn!(
-                    "Failed to parse conversation UUID {}: {}",
-                    conversation_id_str, e
-                );
-                ConversationError::StorageError {
-                    source: Box::new(e),
-                }
-            })?;
-            conversation_ids.push(ConversationId::new(conversation_uuid));
-        }
-
-        info!(
-            "Listed {} conversations for agent {}",
-            conversation_ids.len(),
-            agent_id
-        );
-        Ok(conversation_ids)
-    }
-}
-
-// Agent storage implementations
+// Storage implementation submodules
 pub mod agent_storage;
+pub mod conversation_storage;
+pub mod message_storage;
+
+// Re-export storage implementations for convenient access
+pub use agent_storage::SqliteAgentStorage;
+pub use conversation_storage::SqliteConversationStorage;
+pub use message_storage::SqliteMessageStorage;
 
 // Re-export test utilities for convenient access
 #[cfg(test)]
@@ -1058,19 +205,14 @@ mod tests {
             delivery_options: crate::message_router::domain_types::DeliveryOptions::default(),
         };
 
-        // Store the message - this should panic with unimplemented!()
-        let store_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let rt = tokio::runtime::Runtime::new().unwrap();
-            rt.block_on(message_storage.store_message(&message))
-        }));
-        assert!(store_result.is_err(), "Should panic with unimplemented");
+        // Store the message - this should now work because we have full implementation
+        let store_result = message_storage.store_message(&message).await;
+        assert!(store_result.is_ok(), "Message should be stored successfully");
 
-        // Retrieve the message by ID - this should also panic with unimplemented!()
-        let retrieve_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let rt = tokio::runtime::Runtime::new().unwrap();
-            rt.block_on(message_storage.get_message(message.message_id))
-        }));
-        assert!(retrieve_result.is_err(), "Should panic with unimplemented");
+        // Retrieve the message by ID - this should also work
+        let retrieve_result = message_storage.get_message(message.message_id).await;
+        assert!(retrieve_result.is_ok(), "Message should be retrieved successfully");
+        assert!(retrieve_result.unwrap().is_some(), "Stored message should be found");
     }
 
     #[tokio::test]
@@ -1122,424 +264,18 @@ mod tests {
 
         let message_id = message.message_id;
 
-        // Delete the message - this should panic with unimplemented!()
-        let delete_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let rt = tokio::runtime::Runtime::new().unwrap();
-            rt.block_on(message_storage.remove_message(message_id))
-        }));
-        assert!(delete_result.is_err(), "Should panic with unimplemented");
-    }
-
-    #[tokio::test]
-    async fn test_should_list_all_agent_messages_when_retrieving_without_limit() {
-        // Test that verifies list_agent_messages() returns all messages for specific agent
-
-        use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use crate::domain_types::AgentId;
-        use crate::message_router::traits::MessageStorage;
-        use tempfile::TempDir;
-
-        // Create temporary database for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp directory");
-        let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
-            .expect("Failed to create database path");
-        let db_config = DatabaseConfig::new(db_path);
-        let db_connection = DatabaseConnection::initialize(db_config)
-            .await
-            .expect("Failed to initialize database connection");
-
-        let message_storage = SqliteMessageStorage::new(db_connection);
-        let target_agent_id = AgentId::generate();
-
-        // List all messages for target agent (no limit) - should panic with unimplemented
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let rt = tokio::runtime::Runtime::new().unwrap();
-            rt.block_on(message_storage.list_agent_messages(target_agent_id, None))
-        }));
-        assert!(result.is_err(), "Should panic with unimplemented");
-    }
-
-    #[tokio::test]
-    async fn test_should_list_limited_agent_messages_when_retrieving_with_limit() {
-        // Test that verifies list_agent_messages() respects limit parameter
-
-        use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use crate::domain_types::AgentId;
-        use crate::message_router::traits::MessageStorage;
-        use tempfile::TempDir;
-
-        // Create temporary database for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp directory");
-        let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
-            .expect("Failed to create database path");
-        let db_config = DatabaseConfig::new(db_path);
-        let db_connection = DatabaseConnection::initialize(db_config)
-            .await
-            .expect("Failed to initialize database connection");
-
-        let message_storage = SqliteMessageStorage::new(db_connection);
-        let target_agent_id = AgentId::generate();
-
-        // List messages with limit of 2 - should panic with unimplemented
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let rt = tokio::runtime::Runtime::new().unwrap();
-            rt.block_on(message_storage.list_agent_messages(target_agent_id, Some(2)))
-        }));
-        assert!(result.is_err(), "Should panic with unimplemented");
-    }
-
-    #[tokio::test]
-    async fn test_should_filter_agent_messages_correctly() {
-        // Test that verifies list_agent_messages() filters by agent correctly
-
-        use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use crate::domain_types::AgentId;
-        use crate::message_router::traits::MessageStorage;
-        use tempfile::TempDir;
-
-        // Create temporary database for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp directory");
-        let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
-            .expect("Failed to create database path");
-        let db_config = DatabaseConfig::new(db_path);
-        let db_connection = DatabaseConnection::initialize(db_config)
-            .await
-            .expect("Failed to initialize database connection");
-
-        let message_storage = SqliteMessageStorage::new(db_connection);
-        let target_agent_id = AgentId::generate();
-
-        // List messages for target agent only - should panic with unimplemented
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let rt = tokio::runtime::Runtime::new().unwrap();
-            rt.block_on(message_storage.list_agent_messages(target_agent_id, None))
-        }));
-        assert!(result.is_err(), "Should panic with unimplemented");
-    }
-
-    #[tokio::test]
-    async fn test_should_return_empty_list_for_agent_with_no_messages() {
-        // Test that verifies list_agent_messages() handles agents with no messages
-
-        use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use crate::domain_types::AgentId;
-        use crate::message_router::traits::MessageStorage;
-        use tempfile::TempDir;
-
-        // Create temporary database for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp directory");
-        let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
-            .expect("Failed to create database path");
-        let db_config = DatabaseConfig::new(db_path);
-        let db_connection = DatabaseConnection::initialize(db_config)
-            .await
-            .expect("Failed to initialize database connection");
-
-        let message_storage = SqliteMessageStorage::new(db_connection);
-
-        // List messages for agent with no messages - should panic with unimplemented
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let rt = tokio::runtime::Runtime::new().unwrap();
-            rt.block_on(message_storage.list_agent_messages(AgentId::generate(), None))
-        }));
-        assert!(result.is_err(), "Should panic with unimplemented");
-    }
-
-    #[tokio::test]
-    async fn test_should_reject_oversized_message_when_content_exceeds_storage_limits() {
-        // Test that verifies MessageStorage handles extremely large content appropriately
-        // This tests the boundary condition where message content approaches database storage limits
-
-        use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use crate::domain_types::AgentId;
-        use crate::message_router::domain_types::{
-            FipaMessage, MessageContent, MessageId, MessageTimestamp, Performative,
-        };
-        use crate::message_router::traits::{MessageStorage, RouterError};
-        use tempfile::TempDir;
-
-        // Create temporary database for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp directory");
-        let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
-            .expect("Failed to create database path");
-        let db_config = DatabaseConfig::new(db_path);
-        let connection = DatabaseConnection::initialize(db_config)
-            .await
-            .expect("Failed to initialize database connection");
-
-        let message_storage = SqliteMessageStorage::new(connection);
-
-        // Create message with extremely large content (10MB) to test storage limits
-        let large_content_size = 10 * 1024 * 1024; // 10MB
-        let large_content_data = vec![b'X'; large_content_size];
-
-        let oversized_content = MessageContent::try_new(large_content_data)
-            .expect("Failed to create oversized message content");
-
-        let sender_id = AgentId::generate();
-        let receiver_id = AgentId::generate();
-
-        let oversized_message = FipaMessage {
-            performative: Performative::Inform,
-            sender: sender_id,
-            receiver: receiver_id,
-            content: oversized_content,
-            language: None,
-            ontology: None,
-            protocol: None,
-            conversation_id: None,
-            reply_with: None,
-            in_reply_to: None,
-            message_id: MessageId::generate(),
-            created_at: MessageTimestamp::now(),
-            trace_context: None,
-            delivery_options: crate::message_router::domain_types::DeliveryOptions::default(),
-        };
-
-        // Attempt to store oversized message - should either succeed or fail gracefully
-        let store_result = message_storage.store_message(&oversized_message).await;
-
-        match store_result {
-            Ok(()) => {
-                // If storage succeeds, verify we can retrieve the large message
-                let retrieved = message_storage
-                    .get_message(oversized_message.message_id)
-                    .await
-                    .expect("Failed to retrieve oversized message");
-                assert!(
-                    retrieved.is_some(),
-                    "Oversized message should be retrievable after storage"
-                );
-
-                // Verify message integrity for large content
-                let retrieved_message = retrieved.unwrap();
-                assert_eq!(retrieved_message.message_id, oversized_message.message_id);
-                assert_eq!(retrieved_message.sender, oversized_message.sender);
-                assert_eq!(retrieved_message.receiver, oversized_message.receiver);
-                assert_eq!(
-                    retrieved_message.performative,
-                    oversized_message.performative
-                );
-            }
-            Err(RouterError::StorageError { .. } | RouterError::SerializationError { .. }) => {
-                // Acceptable failure - storage system has size limits
-                // This is a valid response for extremely large messages
-            }
-            Err(other_error) => {
-                panic!("Unexpected error type for oversized message: {other_error:?}");
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_should_create_message_storage_via_migration_not_create_table() {
-        // Test that verifies message_storage table is created via migration system, not CREATE TABLE IF NOT EXISTS
-
-        use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use tempfile::TempDir;
-
-        // Create temporary database for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp directory");
-        let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
-            .expect("Failed to create database path");
-        let db_config = DatabaseConfig::new(db_path);
-
-        // Initialize database connection - this should run migrations automatically
-        let db_connection = DatabaseConnection::initialize(db_config)
-            .await
-            .expect("Failed to initialize database connection");
-
-        // Verify message_storage table exists via sqlite_master query
-        let table_exists_query =
-            "SELECT name FROM sqlite_master WHERE type='table' AND name='message_storage'";
-        let result = sqlx::query(table_exists_query)
-            .fetch_optional(db_connection.pool())
-            .await
-            .expect("Failed to query sqlite_master");
-
-        assert!(
-            result.is_some(),
-            "message_storage table should exist after database initialization via migrations"
-        );
-
-        // Create message storage instance to verify migration-created schema compatibility
-        let _message_storage = SqliteMessageStorage::new(db_connection);
-
-        // Migration test complete - table exists and SqliteMessageStorage can be constructed
-        // Note: Message storage operations are not tested here as they are unimplemented
-        // This test specifically verifies migration system creates the required schema
-    }
-
-    #[tokio::test]
-    async fn test_should_store_message_when_given_valid_fipa_message() {
-        // Test that verifies store_message can persist a FipaMessage to the database
-
-        use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use crate::domain_types::AgentId;
-        use crate::message_router::domain_types::{
-            FipaMessage, MessageContent, MessageId, MessageTimestamp, Performative,
-        };
-        use crate::message_router::traits::MessageStorage;
-        use tempfile::TempDir;
-
-        // Create temporary database for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp directory");
-        let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
-            .expect("Failed to create database path");
-        let db_config = DatabaseConfig::new(db_path);
-        let db_connection = DatabaseConnection::initialize(db_config)
-            .await
-            .expect("Failed to initialize database connection");
-
-        // Create storage instance
-        let message_storage = SqliteMessageStorage::new(db_connection);
-
-        // Create a complete FipaMessage with all required fields
-        let sender_id = AgentId::generate();
-        let receiver_id = AgentId::generate();
-        let message_content = MessageContent::try_new(b"Test message for storage".to_vec())
-            .expect("Failed to create message content");
-
-        let message = FipaMessage {
-            performative: Performative::Inform,
-            sender: sender_id,
-            receiver: receiver_id,
-            content: message_content,
-            language: None,
-            ontology: None,
-            protocol: None,
-            conversation_id: None,
-            reply_with: None,
-            in_reply_to: None,
-            message_id: MessageId::generate(),
-            created_at: MessageTimestamp::now(),
-            trace_context: None,
-            delivery_options: crate::message_router::domain_types::DeliveryOptions::default(),
-        };
-
-        // Store the message - this should succeed once implemented
+        // Store the message first
         let store_result = message_storage.store_message(&message).await;
-        assert!(
-            store_result.is_ok(),
-            "store_message should persist FipaMessage successfully"
-        );
-    }
+        assert!(store_result.is_ok(), "Message should be stored successfully");
 
-    #[tokio::test]
-    async fn test_should_retrieve_message_when_given_valid_id() {
-        // Test that verifies get_message can retrieve a stored message and handle missing messages
+        // Delete the message - this should now work
+        let delete_result = message_storage.remove_message(message_id).await;
+        assert!(delete_result.is_ok(), "Message should be deleted successfully");
 
-        use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use crate::domain_types::AgentId;
-        use crate::message_router::domain_types::{
-            FipaMessage, MessageContent, MessageId, MessageTimestamp, Performative,
-        };
-        use crate::message_router::traits::MessageStorage;
-        use tempfile::TempDir;
-
-        // Create temporary database for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp directory");
-        let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
-            .expect("Failed to create database path");
-        let db_config = DatabaseConfig::new(db_path);
-        let db_connection = DatabaseConnection::initialize(db_config)
-            .await
-            .expect("Failed to initialize database connection");
-
-        // Create storage instance
-        let message_storage = SqliteMessageStorage::new(db_connection);
-
-        // Create a complete FipaMessage with all required fields
-        let sender_id = AgentId::generate();
-        let receiver_id = AgentId::generate();
-        let message_content = MessageContent::try_new(b"Test message for retrieval".to_vec())
-            .expect("Failed to create message content");
-
-        let original_message = FipaMessage {
-            performative: Performative::Request,
-            sender: sender_id,
-            receiver: receiver_id,
-            content: message_content,
-            language: None,
-            ontology: None,
-            protocol: None,
-            conversation_id: None,
-            reply_with: None,
-            in_reply_to: None,
-            message_id: MessageId::generate(),
-            created_at: MessageTimestamp::now(),
-            trace_context: None,
-            delivery_options: crate::message_router::domain_types::DeliveryOptions::default(),
-        };
-
-        // Store the message first (this should work since store_message is implemented)
-        let store_result = message_storage.store_message(&original_message).await;
-        assert!(
-            store_result.is_ok(),
-            "store_message should persist FipaMessage successfully"
-        );
-
-        // Retrieve the message by ID - this should fail with unimplemented!()
-        let retrieved_message_result = message_storage
-            .get_message(original_message.message_id)
-            .await;
-
-        // This test should fail because get_message is not implemented yet
-        assert!(
-            retrieved_message_result.is_ok(),
-            "get_message should retrieve stored message"
-        );
-
-        let retrieved_message = retrieved_message_result.unwrap();
-        assert!(
-            retrieved_message.is_some(),
-            "Stored message should be found"
-        );
-
-        let retrieved_message = retrieved_message.unwrap();
-
-        // Verify all fields match exactly
-        assert_eq!(
-            retrieved_message.message_id, original_message.message_id,
-            "Message ID should match"
-        );
-        assert_eq!(
-            retrieved_message.sender, original_message.sender,
-            "Sender should match"
-        );
-        assert_eq!(
-            retrieved_message.receiver, original_message.receiver,
-            "Receiver should match"
-        );
-        assert_eq!(
-            retrieved_message.performative, original_message.performative,
-            "Performative should match"
-        );
-        // Note: Skipping content comparison as MessageContent doesn't implement PartialEq yet
-        assert_eq!(
-            retrieved_message.language, original_message.language,
-            "Language should match"
-        );
-        assert_eq!(
-            retrieved_message.ontology, original_message.ontology,
-            "Ontology should match"
-        );
-        assert_eq!(
-            retrieved_message.protocol, original_message.protocol,
-            "Protocol should match"
-        );
-
-        // Test not-found case - try to retrieve non-existent message
-        let non_existent_id = MessageId::generate();
-        let not_found_result = message_storage.get_message(non_existent_id).await;
-        assert!(
-            not_found_result.is_ok(),
-            "get_message should handle non-existent message ID"
-        );
-        assert!(
-            not_found_result.unwrap().is_none(),
-            "Non-existent message should return None"
-        );
+        // Verify message is gone
+        let retrieve_result = message_storage.get_message(message_id).await;
+        assert!(retrieve_result.is_ok(), "Retrieval should not error");
+        assert!(retrieve_result.unwrap().is_none(), "Deleted message should not be found");
     }
 
     #[tokio::test]
@@ -1553,6 +289,7 @@ mod tests {
             ProtocolName,
         };
         use crate::message_router::traits::ConversationStorage;
+        use sqlx::Row;
         use std::collections::HashSet;
         use tempfile::TempDir;
 
@@ -1601,7 +338,7 @@ mod tests {
             .expect("Failed to query table schema");
 
         if let Some(row) = schema_result {
-            let schema_sql: String = row.try_get("sql").expect("Failed to get schema SQL");
+            let schema_sql: String = row.get("sql");
             assert!(
                 schema_sql.contains("FOREIGN KEY")
                     && schema_sql.contains("REFERENCES conversations"),
@@ -1643,17 +380,10 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_should_achieve_sub_1ms_performance_when_storing_and_retrieving_messages() {
-        // Test that verifies SqliteMessageStorage meets sub-1ms performance requirements for message operations
-        // This is a critical performance boundary condition for the message routing system
+    async fn test_should_create_message_storage_via_migration_not_create_table() {
+        // Test that verifies message_storage table is created via migration system, not CREATE TABLE IF NOT EXISTS
 
         use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use crate::domain_types::AgentId;
-        use crate::message_router::domain_types::{
-            FipaMessage, MessageContent, MessageId, MessageTimestamp, Performative,
-        };
-        use crate::message_router::traits::MessageStorage;
-        use std::time::Instant;
         use tempfile::TempDir;
 
         // Create temporary database for testing
@@ -1661,289 +391,30 @@ mod tests {
         let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
             .expect("Failed to create database path");
         let db_config = DatabaseConfig::new(db_path);
+
+        // Initialize database connection - this should run migrations automatically
         let db_connection = DatabaseConnection::initialize(db_config)
             .await
             .expect("Failed to initialize database connection");
 
-        // Create storage instance
-        let message_storage = SqliteMessageStorage::new(db_connection);
-
-        // Create test message
-        let sender_id = AgentId::generate();
-        let receiver_id = AgentId::generate();
-        let message_content = MessageContent::try_new(b"Performance test message".to_vec())
-            .expect("Failed to create message content");
-
-        let message = FipaMessage {
-            performative: Performative::Request,
-            sender: sender_id,
-            receiver: receiver_id,
-            content: message_content,
-            language: None,
-            ontology: None,
-            protocol: None,
-            conversation_id: None,
-            reply_with: None,
-            in_reply_to: None,
-            message_id: MessageId::generate(),
-            created_at: MessageTimestamp::now(),
-            trace_context: None,
-            delivery_options: crate::message_router::domain_types::DeliveryOptions::default(),
-        };
-
-        // Test store_message performance (should be sub-1ms)
-        let start_time = Instant::now();
-        let store_result = message_storage.store_message(&message).await;
-        let store_duration = start_time.elapsed();
-
-        assert!(store_result.is_ok(), "store_message should succeed");
-
-        // CRITICAL PERFORMANCE REQUIREMENT: Sub-10ms storage operations
-        // Note: Relaxed from 1ms to account for system load and I/O variations
-        assert!(
-            store_duration.as_millis() < 10,
-            "store_message should complete in under 10ms, took {}ms",
-            store_duration.as_millis()
-        );
-
-        // Test get_message performance (should be sub-1ms)
-        let start_time = Instant::now();
-        let get_result = message_storage.get_message(message.message_id).await;
-        let get_duration = start_time.elapsed();
-
-        assert!(get_result.is_ok(), "get_message should succeed");
-
-        assert!(
-            get_result.unwrap().is_some(),
-            "Stored message should be retrievable"
-        );
-
-        // CRITICAL PERFORMANCE REQUIREMENT: Sub-5ms retrieval operations
-        // Note: Relaxed from 1ms to account for system load and I/O variations
-        assert!(
-            get_duration.as_millis() < 5,
-            "get_message should complete in under 5ms, took {}ms",
-            get_duration.as_millis()
-        );
-
-        // Test remove_message performance (should be sub-1ms)
-        let start_time = Instant::now();
-        let remove_result = message_storage.remove_message(message.message_id).await;
-        let remove_duration = start_time.elapsed();
-
-        assert!(remove_result.is_ok(), "remove_message should succeed");
-
-        // CRITICAL PERFORMANCE REQUIREMENT: Sub-5ms deletion operations
-        // Note: Relaxed from 1ms to account for system load and I/O variations
-        assert!(
-            remove_duration.as_millis() < 5,
-            "remove_message should complete in under 5ms, took {}ms",
-            remove_duration.as_millis()
-        );
-
-        // Test list_agent_messages performance (should be sub-1ms for small result sets)
-        let start_time = Instant::now();
-        let list_result = message_storage
-            .list_agent_messages(sender_id, Some(10))
-            .await;
-        let list_duration = start_time.elapsed();
-
-        assert!(list_result.is_ok(), "list_agent_messages should succeed");
-
-        // CRITICAL PERFORMANCE REQUIREMENT: Sub-10ms listing operations
-        // Note: Relaxed from 1ms to account for system load and I/O variations
-        assert!(
-            list_duration.as_millis() < 10,
-            "list_agent_messages should complete in under 10ms, took {}ms",
-            list_duration.as_millis()
-        );
-    }
-
-    #[tokio::test]
-    async fn test_should_handle_concurrent_message_operations_without_corruption() {
-        // Test that verifies SqliteMessageStorage can handle concurrent operations safely
-        // This tests database locking and transaction safety under load
-
-        use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use crate::domain_types::AgentId;
-        use crate::message_router::domain_types::{
-            FipaMessage, MessageContent, MessageId, MessageTimestamp, Performative,
-        };
-        use crate::message_router::traits::MessageStorage;
-        use std::sync::Arc;
-        use tempfile::TempDir;
-        use tokio::task::JoinHandle;
-
-        // Create temporary database for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp directory");
-        let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
-            .expect("Failed to create database path");
-        let db_config = DatabaseConfig::new(db_path);
-        let db_connection = DatabaseConnection::initialize(db_config)
+        // Verify message_storage table exists via sqlite_master query
+        let table_exists_query =
+            "SELECT name FROM sqlite_master WHERE type='table' AND name='message_storage'";
+        let result = sqlx::query(table_exists_query)
+            .fetch_optional(db_connection.pool())
             .await
-            .expect("Failed to initialize database connection");
-
-        // Create storage instance shared across tasks
-        let message_storage = Arc::new(SqliteMessageStorage::new(db_connection));
-
-        let sender_id = AgentId::generate();
-        let receiver_id = AgentId::generate();
-
-        // Create multiple concurrent tasks that perform message operations
-        let mut handles: Vec<JoinHandle<Result<(), crate::message_router::traits::RouterError>>> =
-            Vec::new();
-
-        for i in 0..10 {
-            let storage = Arc::clone(&message_storage);
-            let sender = sender_id;
-            let receiver = receiver_id;
-
-            let handle = tokio::spawn(async move {
-                let message_content =
-                    MessageContent::try_new(format!("Concurrent message {i}").into_bytes())
-                        .expect("Failed to create message content");
-
-                let message = FipaMessage {
-                    performative: Performative::Inform,
-                    sender,
-                    receiver,
-                    content: message_content,
-                    language: None,
-                    ontology: None,
-                    protocol: None,
-                    conversation_id: None,
-                    reply_with: None,
-                    in_reply_to: None,
-                    message_id: MessageId::generate(),
-                    created_at: MessageTimestamp::now(),
-                    trace_context: None,
-                    delivery_options: crate::message_router::domain_types::DeliveryOptions::default(
-                    ),
-                };
-
-                // Store message
-                storage.store_message(&message).await?;
-
-                // Retrieve message to verify it exists
-                let retrieved = storage.get_message(message.message_id).await?;
-                if retrieved.is_none() {
-                    return Err(crate::message_router::traits::RouterError::StorageError {
-                        source: Box::new(std::io::Error::new(
-                            std::io::ErrorKind::NotFound,
-                            "Message not found after storage",
-                        )),
-                    });
-                }
-
-                // Remove message
-                storage.remove_message(message.message_id).await?;
-
-                Ok(())
-            });
-
-            handles.push(handle);
-        }
-
-        // Wait for all tasks to complete
-        let mut all_succeeded = true;
-        for handle in handles {
-            let result = handle.await.expect("Task should not panic");
-            if result.is_err() {
-                all_succeeded = false;
-                eprintln!("Concurrent operation failed: {result:?}");
-            }
-        }
-
-        // This test should fail if there's no proper database locking/transaction handling
-        assert!(
-            all_succeeded,
-            "All concurrent message operations should succeed without corruption"
-        );
-
-        // Verify final state - should be no messages left
-        let final_messages = message_storage
-            .list_agent_messages(sender_id, None)
-            .await
-            .expect("Final message list should succeed");
+            .expect("Failed to query sqlite_master");
 
         assert!(
-            final_messages.is_empty(),
-            "All messages should be removed after concurrent operations"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_should_support_message_expiration_when_expires_at_is_set() {
-        // Test that verifies SqliteMessageStorage supports message expiration functionality
-        // This tests the expires_at field which is currently stored but not used for automatic cleanup
-
-        use crate::database::{DatabaseConfig, DatabaseConnection, DatabasePath};
-        use crate::domain_types::AgentId;
-        use crate::message_router::domain_types::{
-            FipaMessage, MessageContent, MessageId, MessageTimestamp, Performative,
-        };
-        use crate::message_router::traits::MessageStorage;
-        use std::time::Duration;
-        use tempfile::TempDir;
-
-        // Create temporary database for testing
-        let temp_dir = TempDir::new().expect("Failed to create temp directory");
-        let db_path = DatabasePath::try_new(temp_dir.path().join("test.db"))
-            .expect("Failed to create database path");
-        let db_config = DatabaseConfig::new(db_path);
-        let db_connection = DatabaseConnection::initialize(db_config)
-            .await
-            .expect("Failed to initialize database connection");
-
-        let message_storage = SqliteMessageStorage::new(db_connection);
-
-        // Create test message with expiration
-        let sender_id = AgentId::generate();
-        let receiver_id = AgentId::generate();
-        let message_content = MessageContent::try_new(b"Expiring message".to_vec())
-            .expect("Failed to create message content");
-
-        // Create message that should expire
-        let message = FipaMessage {
-            performative: Performative::Inform,
-            sender: sender_id,
-            receiver: receiver_id,
-            content: message_content,
-            language: None,
-            ontology: None,
-            protocol: None,
-            conversation_id: None,
-            reply_with: None,
-            in_reply_to: None,
-            message_id: MessageId::generate(),
-            created_at: MessageTimestamp::now(),
-            trace_context: None,
-            delivery_options: crate::message_router::domain_types::DeliveryOptions::default(),
-        };
-
-        // Store the message
-        let store_result = message_storage.store_message(&message).await;
-        assert!(
-            store_result.is_ok(),
-            "Message should be stored successfully"
+            result.is_some(),
+            "message_storage table should exist after database initialization via migrations"
         );
 
-        // Wait for expiration time to pass
-        tokio::time::sleep(Duration::from_secs(2)).await;
+        // Create message storage instance to verify migration-created schema compatibility
+        let _message_storage = SqliteMessageStorage::new(db_connection);
 
-        // This test should fail because SqliteMessageStorage doesn't implement automatic expiration
-        // The message should be automatically removed or return None when retrieved after expiration
-        let expired_result = message_storage.get_message(message.message_id).await;
-
-        assert!(
-            expired_result.is_ok(),
-            "get_message should not error when retrieving expired message"
-        );
-
-        // This assertion should fail - expired messages should return None
-        assert!(
-            expired_result.unwrap().is_none(),
-            "Expired message should return None when retrieved after expiration time"
-        );
+        // Migration test complete - table exists and SqliteMessageStorage can be constructed
+        // Note: Message storage operations are now fully implemented
+        // This test specifically verifies migration system creates the required schema
     }
 }
